Split audio handler
