+++ lib/components/player/common/volume_slider_panel.dart
@@ -3,7 +3,7 @@ import 'dart:math' as math;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:yaabsa/util/globals.dart';
-import 'package:yaabsa/util/handler/bg_audio_handler.dart';
+import 'package:yaabsa/util/audio_handler/bg_audio_handler.dart';
 
 class VolumeSliderPanel extends StatelessWidget {
   const VolumeSliderPanel({super.key, required this.axis});
