+++ lib/util/audio_handler/auto/bg_audio_handler_android_auto_ids.dart
@@ -1,4 +1,4 @@
-part of 'bg_audio_handler.dart';
+part of '../bg_audio_handler.dart';
 
 extension _BGAudioHandlerAndroidAutoIds on BGAudioHandler {
   _AndroidAutoPlaybackTarget? _androidAutoPlaybackTargetFromMediaId(String mediaId) {
