+++ lib/util/init.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:audio_service/audio_service.dart';
 import 'package:yaabsa/util/globals.dart';
 import 'package:yaabsa/util/device_capabilities.dart';
-import 'package:yaabsa/util/handler/bg_audio_handler.dart';
+import 'package:yaabsa/util/audio_handler/bg_audio_handler.dart';
 import 'package:yaabsa/util/handler/download_handler.dart';
 import 'package:yaabsa/util/handler/shake_handler.dart';
 import 'package:yaabsa/util/handler/sleep_timer_handler.dart';
